Wrap logsEventSource in a class
