+++ src/commands/logs.ts
@@ -1,7 +1,7 @@
 import { flags as oclifFlags } from '@oclif/command'
 
 import { CustomCommand } from '../oclif/CustomCommand'
-import appsLogs from '../modules/apps/logs'
+import appsLogs from '../modules/apps/logs/logs'
 
 export default class Logs extends CustomCommand {
   static description = 'Show logs of an app on the current directory or a specified one'
